requirements.txt
crewai>=0.28.2
pysbd>=0.3.4  # Fixed regex warnings
pydantic>=2.6.0
openai>=1.12.0
google-search-results>=2.4.2  # For Serper API
python-dotenv>=1.0.0  # For API key management